app/requirements.txt
wheel
Adafruit-Blinka
adafruit-circuitpython-bme680
adafruit-circuitpython-busdevice
firebase-admin
geocoder
google-api-core
google-api-python-client
google-auth
grpcio
RPi.GPIO==0.7.1a4
urllib3
typer
ipython
python-crontab